requirements.txt
fastapi[standard]==0.115.4
uvicorn[standard]
mysql-connector-python
sqlalchemy~=2.0.36
python-dotenv~=1.0.1
pydantic~=2.9.2
pymysql
pyopenssl
python-multipart
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4
starlette~=0.41.2
websockets
pytest